// src/enums/billingEnum.ts
export enum PlanPeriodUnit {
  Year = 'year',
  Month = 'month',
  Day = 'day',
  Hour = 'hour',
  Minute = 'minute',
  Unlimited = 'unlimited',
}

export enum PlanType {
  Beta = 'beta',
  Official = 'official',
  Pioneer = 'pioneer',
  Activity = 'activity',
}

export enum PlanLevel {
  Free = 'free',
  Professional = 'professional',
  Business = 'business',
  Custom = 'custom',
  Pioneer = 'pioneer',
  Enterprise = 'enterprise',
}